+++ src/SwashbucklerDiary.Gtk/BlazorWebView/Gtk/RootComponent.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using GtkSharp.BlazorWebKit;
 
 namespace Microsoft.AspNetCore.Components.WebView.Gtk
 {
